requirements.txt
bcrypt==3.2.2
cffi==1.15.0
click==8.0.4
dnspython==2.2.1
email-validator==1.2.1
Flask==2.0.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.1
Flask-MySQLdb==1.0.1
Flask-WTF==1.0.1
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
mysql-connector-python==8.0.28
mysqlclient==2.1.0
protobuf==3.19.4
pycparser==2.21
PyYAML==6.0
Werkzeug==2.0.3
WTForms==3.0.1
